v3.0.1.0 - renamed getFormattedNmsDouble -> getFormattedNmsInteger (had no reason to be a double, plus floating point might've been an issue in the future) - Added KamiCommon commands, and a test suite for NMS

+++ spigot-nms/src/main/java/com/kamikazejam/kamicommon/nms/provider/ChatColorProvider.java
@@ -7,7 +7,7 @@ import org.jetbrains.annotations.NotNull;
 
 public class ChatColorProvider extends Provider<IChatColorNMS> {
     @Override
-    protected @NotNull IChatColorNMS provide(double ver) {
+    protected @NotNull IChatColorNMS provide(int ver) {
         // If we are 1.16.2+ we can use the getColor() method introduced when hex codes were added
         if (ver >= 1162) {
             return new ChatColor1_16_R2();
